refactor(mbx): split session responsibilities
