+++ crates/mbx/src/session.rs
@@ -3,30 +3,54 @@
 
 use crate::config::Config;
 use crate::events::{ActionDetail, ActionOutcome, EventWriter};
-use crate::util::{duration_ns, format_duration, write_atomic};
-use bytesize::ByteSize;
-use eyre::{Context, Result, bail};
+use crate::util::duration_ns;
+use eyre::Result;
 use log::{debug, warn};
 use mbx_cache_cc::CcLanguage;
+#[cfg(test)]
+use mbx_cache_core::AGENT_PROTOCOL_VERSION;
 use mbx_cache_core::{
-    AGENT_PROTOCOL_VERSION, AgentEvent, AgentEventObserver, AgentRemoteCache, AgentRequest,
-    AgentResponse, AgentStats, CacheAgent, CacheDigest, FileDigestCache, FileDigestScope,
-    FileIdentity, NoFileDigestCache, RecordedFileDigest, canonical_json,
+    AgentEvent, AgentEventObserver, AgentRemoteCache, AgentRequest, AgentResponse, AgentStats,
+    CacheAgent, CacheDigest, FileDigestCache, FileDigestScope, FileIdentity, NoFileDigestCache,
+    RecordedFileDigest, canonical_json,
 };
 use serde::Serialize;
-use std::cell::RefCell;
 use std::collections::BTreeMap;
 use std::ffi::{OsStr, OsString};
-#[cfg(unix)]
-use std::io::{BufRead, BufReader, Write};
 use std::path::{Path, PathBuf};
 use std::process::{Command, ExitCode};
 #[cfg(unix)]
-use std::sync::atomic::{AtomicU64, Ordering};
+use std::sync::atomic::AtomicU64;
 use std::sync::{Arc, Mutex};
 use std::time::Instant;
 use tokio::sync::oneshot;
 use tokio::task::JoinHandle;
+
+mod client;
+mod diagnostics;
+mod server;
+mod shims;
+mod stats;
+
+#[cfg(test)]
+use client::validate_handshake_response;
+use client::{request_agent_at, request_standalone_agent};
+pub(crate) use diagnostics::{note, report_shim_warning, reserve_stderr_for_compiler};
+use server::spawn_server;
+use shims::{CC_CRATE_ENV, CcShims, install_cc_shims, install_session_shim};
+pub use shims::{
+    PathShims, ShimLink, install_path_shims, install_shim, install_shim_named, shim_file_name,
+};
+#[cfg(test)]
+use shims::{
+    TargetedCompiler, resolve_in_path, resolve_named_compiler, resolve_on_path, symlink_shim,
+    targeted_compiler_language,
+};
+use stats::StatsReport;
+pub use stats::display_stats;
+pub(crate) use stats::session_was_active;
+#[cfg(test)]
+use stats::{cache_misses, should_display_stats, stale_manifest_note, write_stats_report};
 
 pub const RUSTC_SHIM_STEM: &str = "mbx-rustc";
 pub const CC_SHIM_STEM: &str = "mbx-cc";
@@ -589,1151 +613,6 @@ fn action_remote_cache(config: &Config, store: &Path) -> Result<Option<AgentRemo
         mode,
         staging_dir: store.join("remote"),
     }))
-}
-
-#[derive(Serialize)]
-struct StatsReport {
-    version: u8,
-    session_duration_ns: u64,
-    lookups: u64,
-    hits: u64,
-    misses: u64,
-    unconsulted: u64,
-    compiler_invocations_avoided: u64,
-    estimated_compiler_duration_avoided_ns: u64,
-    compiler: BTreeMap<String, CompilerStatsReport>,
-    slow_compilations: Vec<SlowCompilationReport>,
-    verifications: u64,
-    divergences: u64,
-    prefetched_actions: u64,
-    predictions_loaded: u64,
-    prefetch_runs: u64,
-    bypasses: BTreeMap<String, u64>,
-    downloaded_bytes: u64,
-    uploaded_bytes: u64,
-    background_uploads: u64,
-    background_upload_failures: u64,
-    remote_blob_pack_uploads: u64,
-    remote_blob_pack_upload_blobs: u64,
-    upload_drain_duration_ns: u64,
-    stored_bytes: u64,
-    restored_output_files: u64,
-    restored_output_bytes: u64,
-    reflinked_output_files: u64,
-    reflinked_output_bytes: u64,
-    copied_output_files: u64,
-    copied_output_bytes: u64,
-    reused_output_files: u64,
-    reused_output_bytes: u64,
-    remote_failures: u64,
-    remote_manifest_lookups: u64,
-    remote_action_lookups: u64,
-    remote_blob_requests: u64,
-    remote_blob_pack_requests: u64,
-    remote_blob_pack_blobs: u64,
-    remote_manifest_lookup_duration_ns: u64,
-    remote_action_lookup_duration_ns: u64,
-    remote_blob_transfer_duration_ns: u64,
-    local_cas_write_duration_ns: u64,
-    prefetch_duration_ns: u64,
-    materialization_duration_ns: u64,
-}
-
-#[derive(Serialize)]
-struct CompilerStatsReport {
-    invocations: u64,
-    duration_ns: u64,
-}
-
-#[derive(Serialize)]
-struct SlowCompilationReport {
-    crate_name: String,
-    duration_ns: u64,
-}
-
-impl From<&AgentStats> for StatsReport {
-    fn from(stats: &AgentStats) -> Self {
-        Self {
-            version: 4,
-            session_duration_ns: stats.session_duration_ns,
-            lookups: stats.lookups,
-            hits: stats.hits,
-            misses: cache_misses(stats),
-            unconsulted: stats.unconsulted,
-            compiler_invocations_avoided: stats.hits,
-            estimated_compiler_duration_avoided_ns: stats.avoided_compiler_duration_ns,
-            compiler: stats
-                .compiler
-                .iter()
-                .map(|(outcome, stats)| {
-                    (
-                        outcome.clone(),
-                        CompilerStatsReport {
-                            invocations: stats.invocations,
-                            duration_ns: stats.duration_ns,
-                        },
-                    )
-                })
-                .collect(),
-            slow_compilations: slow_compilations(stats)
-                .into_iter()
-                .map(|(crate_name, duration_ns)| SlowCompilationReport {
-                    crate_name: crate_name.clone(),
-                    duration_ns: *duration_ns,
-                })
-                .collect(),
-            verifications: stats.verifications,
-            divergences: stats.divergences,
-            prefetched_actions: stats.prefetched_actions,
-            predictions_loaded: stats.predictions_loaded,
-            prefetch_runs: stats.prefetch_runs,
-            bypasses: stats.bypasses.clone(),
-            downloaded_bytes: stats.downloaded_bytes,
-            uploaded_bytes: stats.uploaded_bytes,
-            background_uploads: stats.background_uploads,
-            background_upload_failures: stats.background_upload_failures,
-            remote_blob_pack_uploads: stats.remote_blob_pack_uploads,
-            remote_blob_pack_upload_blobs: stats.remote_blob_pack_upload_blobs,
-            upload_drain_duration_ns: stats.upload_drain_duration_ns,
-            stored_bytes: stats.stored_bytes,
-            restored_output_files: stats.restored_output_files,
-            restored_output_bytes: stats.restored_output_bytes,
-            reflinked_output_files: stats.reflinked_output_files,
-            reflinked_output_bytes: stats.reflinked_output_bytes,
-            copied_output_files: stats.copied_output_files,
-            copied_output_bytes: stats.copied_output_bytes,
-            reused_output_files: stats.reused_output_files,
-            reused_output_bytes: stats.reused_output_bytes,
-            remote_failures: stats.remote_failures,
-            remote_manifest_lookups: stats.remote_manifest_lookups,
-            remote_action_lookups: stats.remote_action_lookups,
-            remote_blob_requests: stats.remote_blob_requests,
-            remote_blob_pack_requests: stats.remote_blob_pack_requests,
-            remote_blob_pack_blobs: stats.remote_blob_pack_blobs,
-            remote_manifest_lookup_duration_ns: stats.remote_manifest_lookup_duration_ns,
-            remote_action_lookup_duration_ns: stats.remote_action_lookup_duration_ns,
-            remote_blob_transfer_duration_ns: stats.remote_blob_transfer_duration_ns,
-            local_cas_write_duration_ns: stats.local_cas_write_duration_ns,
-            prefetch_duration_ns: stats.prefetch_duration_ns,
-            materialization_duration_ns: stats.materialization_duration_ns,
-        }
-    }
-}
-
-/// Report a finished session to stderr, and to a JSON file when configured.
-pub fn display_stats(stats: &AgentStats, config: &Config) {
-    if let Some(path) = &config.stats_report
-        && let Err(error) = write_stats_report(path, stats)
-    {
-        warn!(
-            "the statistics report could not be written to {}: {error}",
-            path.display()
-        );
-    }
-    if !should_display_stats(stats) {
-        return;
-    }
-    note(&format!(
-        "mbx[cache]: {} hits, {} misses, {} prefetched; {} downloaded, {} uploaded, {} stored locally",
-        stats.hits,
-        cache_misses(stats),
-        stats.prefetched_actions,
-        ByteSize::b(stats.downloaded_bytes).display().iec(),
-        ByteSize::b(stats.uploaded_bytes).display().iec(),
-        ByteSize::b(stats.stored_bytes).display().iec(),
-    ));
-    if stats.remote_failures > 0 {
-        // A remote cache that fails every request reports the same hits, misses
-        // and bytes as one that was simply empty, and the warnings explaining
-        // why scroll past hundreds of lines earlier. Without this line the
-        // summary reads as "the remote had nothing for us" no matter which it
-        // was, and a cache that has stopped working looks like a cold one.
-        note(&format!(
-            "mbx[cache]: the remote cache failed {} of its requests; this build ran without what it could not reach, and the warnings above say why",
-            stats.remote_failures,
-        ));
-    }
-    if stats.unconsulted > 0 {
-        // A cold target directory hits this for everything it compiles, and
-        // reporting only hits and misses there says "0 hits, 0 misses" -- which
-        // reads as though the cache was asked and had nothing, rather than that
-        // it was never asked. These compilations are still stored afterwards.
-        note(&format!(
-            "mbx[cache]: could not look up {} compilations: no usable dep-info from an earlier build and no prediction to derive an action key from",
-            stats.unconsulted,
-        ));
-        if let Some(explanation) = stale_manifest_note(stats) {
-            note(&explanation);
-        }
-    }
-    if !stats.bypasses.is_empty() {
-        let total: u64 = stats.bypasses.values().sum();
-        // Most frequent first: the head of this list is where the next win is.
-        let mut reasons: Vec<_> = stats.bypasses.iter().collect();
-        reasons.sort_by(|left, right| right.1.cmp(left.1).then(left.0.cmp(right.0)));
-        let detail = reasons
-            .iter()
-            .map(|(kind, count)| format!("{count} {kind}"))
-            .collect::<Vec<_>>()
-            .join(", ");
-        note(&format!(
-            "mbx[cache]: bypassed {total} compilations: {detail}"
-        ));
-    }
-    if !stats.compiler.is_empty() || stats.avoided_compiler_duration_ns > 0 {
-        let spent = stats.compiler.values().fold(0_u64, |total, compiler| {
-            total.saturating_add(compiler.duration_ns)
-        });
-        let detail = stats
-            .compiler
-            .iter()
-            .map(|(outcome, compiler)| {
-                format!(
-                    "{} {} in {}",
-                    compiler.invocations,
-                    outcome,
-                    format_nanos(compiler.duration_ns)
-                )
-            })
-            .collect::<Vec<_>>()
-            .join(", ");
-        note(&format!(
-            "mbx[cache]: compiler time: {} estimated avoided; {} spent ({detail})",
-            format_nanos(stats.avoided_compiler_duration_ns),
-            format_nanos(spent),
-        ));
-        if let Some(compiler) = stats.compiler.get("incremental") {
-            // The compiler-time line above already counts these; what it cannot
-            // say is why they are absent from the store.
-            note(&format!(
-                "mbx[cache]: {} compilations kept their own incremental state, so they were not stored",
-                compiler.invocations
-            ));
-        }
-        let slow = slow_compilations(stats);
-        if !slow.is_empty() {
-            note(&format!(
-                "mbx[cache]: slowest uncached crates: {}",
-                slow.into_iter()
-                    .map(|(name, duration)| format!("{name} {}", format_nanos(*duration)))
-                    .collect::<Vec<_>>()
-                    .join(", ")
-            ));
-        }
-    }
-    let remote_lookup_duration_ns = stats
-        .remote_manifest_lookup_duration_ns
-        .saturating_add(stats.remote_action_lookup_duration_ns);
-    note(&format!(
-        "mbx[cache]: timing: {} session, {} prefetch; cumulative {} remote lookup, {} blob transfer, {} CAS write, {} materialization",
-        format_nanos(stats.session_duration_ns),
-        format_nanos(stats.prefetch_duration_ns),
-        format_nanos(remote_lookup_duration_ns),
-        format_nanos(stats.remote_blob_transfer_duration_ns),
-        format_nanos(stats.local_cas_write_duration_ns),
-        format_nanos(stats.materialization_duration_ns),
-    ));
-    if stats.background_uploads > 0 || stats.background_upload_failures > 0 {
-        let packed = if stats.remote_blob_pack_uploads > 0 {
-            format!(
-                " ({} of them in {} packs)",
-                stats.remote_blob_pack_upload_blobs, stats.remote_blob_pack_uploads
-            )
-        } else {
-            String::new()
-        };
-        note(&format!(
-            "mbx[cache]: uploads: {} published{packed}, {} not published; {} waited for after the build",
-            stats.background_uploads,
-            stats.background_upload_failures,
-            format_nanos(stats.upload_drain_duration_ns),
-        ));
-    }
-    if stats.restored_output_files > 0 {
-        note(&format!(
-            "mbx[cache]: materialization: {} outputs ({}) reflinked, {} outputs ({}) copied, {} outputs ({}) already in place",
-            stats.reflinked_output_files,
-            ByteSize::b(stats.reflinked_output_bytes).display().iec(),
-            stats.copied_output_files,
-            ByteSize::b(stats.copied_output_bytes).display().iec(),
-            stats.reused_output_files,
-            ByteSize::b(stats.reused_output_bytes).display().iec(),
-        ));
-    }
-    if stats.verifications > 0 {
-        note(&format!(
-            "mbx[cache]: qualification: {} verified, {} diverged",
-            stats.verifications, stats.divergences,
-        ));
-    }
-}
-
-/// Explain a session that loaded a full manifest and matched none of it.
-///
-/// "No prediction to derive an action key from" reads as an empty store, and
-/// for a warm store that just watched its compiler change underneath it -- a
-/// CI runner image updating its preinstalled toolchain is the common way --
-/// that reading sends people hunting for restore failures. The distinction is
-/// observable: predictions were loaded, and not one lookup was ever made.
-fn stale_manifest_note(stats: &AgentStats) -> Option<String> {
-    (stats.unconsulted > 0 && stats.lookups == 0 && stats.predictions_loaded > 0).then(|| {
-        format!(
-            "mbx[cache]: a manifest predicting {} compilations was loaded, but none matched this build; the compiler or its flags changed since they were recorded (a toolchain update does this)",
-            stats.predictions_loaded,
-        )
-    })
-}
-
-/// Whether the cache took part in this build at all.
-///
-/// A run that never consulted or stored anything -- `cargo --help`, a build
-/// cargo declined -- has nothing to report and should not be counted as a
-/// build in the lifetime totals.
-pub(crate) fn session_was_active(stats: &AgentStats) -> bool {
-    should_display_stats(stats)
-}
-
-fn should_display_stats(stats: &AgentStats) -> bool {
-    stats.lookups > 0
-        || stats.unconsulted > 0
-        || stats.stores > 0
-        || stats.verifications > 0
-        || stats.downloaded_bytes > 0
-        || stats.uploaded_bytes > 0
-        // An action result published on its own moves no payload bytes, and is
-        // still the whole of what a session did.
-        || stats.background_uploads > 0
-        || !stats.bypasses.is_empty()
-        || !stats.compiler.is_empty()
-        || stats.avoided_compiler_duration_ns > 0
-        // A session that reached a remote cache and got nothing but failures has
-        // nothing else to report, and is the session most worth reporting.
-        || stats.remote_failures > 0
-}
-
-/// Write to stderr without failing the build when the pipe is closed.
-pub(crate) fn note(message: &str) {
-    use std::io::Write as _;
-    let _ = writeln!(std::io::stderr(), "{message}");
-}
-
-/// Whether this process's stderr belongs to the compiler it stands in for.
-///
-/// Set once at cc-shim entry. Build scripts read an intercepted compiler's
-/// stderr as part of its answer -- cc-rs marks a probed flag unsupported the
-/// moment anything lands there -- so one printed warning changes the flags of
-/// every compilation the build script produces afterwards, and with them
-/// every action key, orphaning the whole build's predictions.
-static STDERR_RESERVED: std::sync::atomic::AtomicBool = std::sync::atomic::AtomicBool::new(false);
-
-/// Declare that this process replays compiler output on stderr and must not
-/// mix its own diagnostics into it.
-pub(crate) fn reserve_stderr_for_compiler() {
-    STDERR_RESERVED.store(true, std::sync::atomic::Ordering::Relaxed);
-}
-
-/// Report a shim diagnostic without polluting a reserved stderr.
-///
-/// Delivered to the session's agent, which prints it from the process that
-/// owns the build. When the agent cannot take it, the message falls back to
-/// this process's stderr only where that stream is not the compiler's --
-/// losing a diagnostic costs a little visibility, while poisoning a configure
-/// probe costs the build its cache keys.
-pub(crate) fn report_shim_warning(message: &str) {
-    let mut message = message.replace(['\n', '\r'], "; ");
-    // Stay under the agent's acceptance limit rather than losing the whole
-    // diagnostic to it; the start of an error chain names the failure.
-    if message.len() > 2048 {
-        let end = (0..=2048).rfind(|&index| message.is_char_boundary(index));
-        message.truncate(end.unwrap_or_default());
-        message.push_str("...");
-    }
-    let delivered = matches!(
-        request_agent(&[AgentRequest::RecordWarning {
-            message: message.clone(),
-        }])
-        .map(|responses| responses.into_iter().next()),
-        Ok(Some(AgentResponse::WarningRecorded))
-    );
-    if !delivered && !STDERR_RESERVED.load(std::sync::atomic::Ordering::Relaxed) {
-        note(&format!("mbx[warning]: {message}"));
-    }
-}
-
-fn write_stats_report(path: &Path, stats: &AgentStats) -> Result<()> {
-    let mut report = serde_json::to_vec_pretty(&StatsReport::from(stats))?;
-    report.push(b'\n');
-    write_atomic(path, &report)
-}
-
-fn format_nanos(nanoseconds: u64) -> String {
-    format_duration(std::time::Duration::from_nanos(nanoseconds))
-}
-
-fn slow_compilations(stats: &AgentStats) -> Vec<(&String, &u64)> {
-    let mut slow = stats.slow_compilations.iter().collect::<Vec<_>>();
-    slow.sort_by(|left, right| right.1.cmp(left.1).then(left.0.cmp(right.0)));
-    slow.truncate(5);
-    slow
-}
-
-fn cache_misses(stats: &AgentStats) -> u64 {
-    stats
-        .lookups
-        .saturating_sub(stats.hits)
-        .saturating_sub(stats.verifications)
-}
-
-/// Installed C and C++ shims, and the compilers they stand in for.
-struct CcShims {
-    /// Installed C shim and the compiler it stands in for, when there is one.
-    cc: Option<(PathBuf, PathBuf)>,
-    /// The same for C++.
-    cxx: Option<(PathBuf, PathBuf)>,
-    /// Target-specific compilers the build named, each behind its own shim.
-    targeted: Vec<TargetedCompiler>,
-}
-
-/// A compiler a cross build asked for by name, and the shim standing in for it.
-struct TargetedCompiler {
-    /// The `cc` crate variable that named it, such as `CC_aarch64-linux-musl`.
-    variable: String,
-    /// File name of the shim installed for it, which is also its pin key.
-    shim_name: String,
-    /// Absolute path to the shim.
-    shim: PathBuf,
-    /// The compiler the build chose, which the shim execs.
-    real: PathBuf,
-}
-
-impl CcShims {
-    /// Point build scripts at whichever shims were installed.
-    ///
-    /// A language with no compiler on the machine contributes nothing, so a C
-    /// build on an image without a C++ compiler still gets its caching.
-    fn apply_host(&self, environment: &mut BTreeMap<String, String>) {
-        for (installed, host, real) in [
-            (&self.cc, "HOST_CC", REAL_CC_ENV),
-            (&self.cxx, "HOST_CXX", REAL_CXX_ENV),
-        ] {
-            if let Some((shim, compiler)) = installed {
-                environment.insert(host.into(), shim.to_string_lossy().into_owned());
-                environment.insert(real.into(), compiler.to_string_lossy().into_owned());
-            }
-        }
-    }
-
-    /// Point each variable that named a cross compiler at its own shim.
-    fn apply_targeted(&self, environment: &mut BTreeMap<String, String>) {
-        for targeted in &self.targeted {
-            environment.insert(
-                targeted.variable.clone(),
-                targeted.shim.to_string_lossy().into_owned(),
-            );
-        }
-    }
-
-    /// Pins for the shims that stand in for a named cross compiler.
-    ///
-    /// They share the map the standalone shims use, which is what lets one
-    /// shim binary serve several compilers: it looks itself up by the name it
-    /// was invoked under.
-    fn pins(&self) -> BTreeMap<String, PathBuf> {
-        self.targeted
-            .iter()
-            .map(|targeted| (targeted.shim_name.clone(), targeted.real.clone()))
-            .collect()
-    }
-}
-
-/// Whether an environment variable is how the `cc` crate names a compiler for
-/// a particular target.
-///
-/// `TARGET_CC` and `TARGET_CXX` apply to whatever the build is cross-compiling
-/// for; `CC_<target>` and `CXX_<target>` name one triple outright.
-fn targeted_compiler_language(variable: &str) -> Option<CcLanguage> {
-    match variable {
-        "TARGET_CC" => return Some(CcLanguage::C),
-        "TARGET_CXX" => return Some(CcLanguage::Cxx),
-        _ => {}
-    }
-    // `CXX_` first: `CC_` is not a prefix of it, but reading it the other way
-    // round invites the mistake.
-    for (prefix, language) in [("CXX_", CcLanguage::Cxx), ("CC_", CcLanguage::C)] {
-        if let Some(target) = variable.strip_prefix(prefix)
-            && is_target_triple(target)
-        {
-            return Some(language);
-        }
-    }
-    None
-}
-
-/// Whether a variable suffix spells a target triple rather than something else
-/// that happens to start with `CC_`.
-///
-/// The `cc` crate hangs its own controls off that prefix -- `CC_FORCE_DISABLE`,
-/// `CC_KNOWN_WRAPPER_CUSTOM`, `CC_ENABLE_DEBUG_OUTPUT` -- and autotools adds
-/// `CC_FOR_BUILD`. Redirecting one of those would not miss a cache hit, it
-/// would answer a question the build asked with a compiler path.
-///
-/// Case is what separates them: a target triple is lowercase and those knobs
-/// are not. A triple also always has at least two components, which a bare word
-/// does not.
-fn is_target_triple(suffix: &str) -> bool {
-    !suffix.is_empty()
-        && suffix.contains(['-', '_'])
-        && suffix.bytes().all(|byte| {
-            byte.is_ascii_lowercase() || byte.is_ascii_digit() || matches!(byte, b'-' | b'_' | b'.')
-        })
-}
-
-/// Variables the `cc` crate consults before falling back to the platform
-/// default. A build that sets any of them has chosen its own compiler, and mbx
-/// stands aside rather than redirecting it.
-const CC_CRATE_ENV: &[&str] = &["CC", "CXX", "HOST_CC", "HOST_CXX"];
-
-/// Install the C and C++ shims, resolving the compilers they will run.
-///
-/// Resolution happens here rather than in the shim so the whole build agrees on
-/// one compiler, and so a machine with no C compiler simply gets no shims
-/// instead of a build script that fails differently than it would have.
-fn install_cc_shims(shims_dir: &Path) -> Result<Option<CcShims>> {
-    if !cfg!(unix) {
-        return Ok(None);
-    }
-    let executable = std::env::current_exe().wrap_err("failed to locate the running mbx binary")?;
-    // Each language stands alone. An image with a C compiler and no C++ one is
-    // ordinary, and it must not cost a C-only sys-crate its caching.
-    let real_cc = resolve_on_path(CcLanguage::C.default_driver());
-    let real_cxx = resolve_on_path(CcLanguage::Cxx.default_driver());
-    // Wrapped first, because a cross image is entitled to ship the driver it
-    // cross-compiles with and no host `cc` at all. Deciding there is nothing to
-    // do before looking would leave exactly that build uncached.
-    std::fs::create_dir_all(shims_dir)?;
-    let targeted = wrap_targeted_compilers(&executable, shims_dir)?;
-    if real_cc.is_none() && real_cxx.is_none() && targeted.is_empty() {
-        debug!("no C or C++ compiler was found on PATH; build script compiles are not cached");
-        return Ok(None);
-    }
-    let shim = |language: CcLanguage| -> Result<PathBuf> {
-        let destination = shims_dir.join(shim_file_name(language.shim_stem()));
-        link_path_shim(&executable, &destination)?;
-        Ok(destination)
-    };
-    let mut installed = CcShims {
-        cc: None,
-        cxx: None,
-        targeted,
-    };
-    if let Some(real) = real_cc {
-        installed.cc = Some((shim(CcLanguage::C)?, real));
-    }
-    if let Some(real) = real_cxx {
-        installed.cxx = Some((shim(CcLanguage::Cxx)?, real));
-    }
-    Ok(Some(installed))
-}
-
-/// Put a shim in front of every cross compiler the build named for itself.
-///
-/// Deriving one is not an option: which compiler a target implies lives in the
-/// `cc` crate's own tables, and guessing wrong would not cost a cache hit, it
-/// would build the object with the wrong compiler. So only a compiler the build
-/// asked for by name is wrapped, and only when the name resolves to a single
-/// executable -- a value like `ccache gcc` is a command, not a path, and is
-/// left alone.
-fn wrap_targeted_compilers(executable: &Path, shims_dir: &Path) -> Result<Vec<TargetedCompiler>> {
-    let mut wrapped = Vec::new();
-    for (variable, value) in std::env::vars() {
-        let Some(language) = targeted_compiler_language(&variable) else {
-            continue;
-        };
-        let Some(real) = resolve_named_compiler(&value, executable, shims_dir) else {
-            debug!("{variable} does not name a single executable; it is left as it is");
-            continue;
-        };
-        // Named for the variable so one shim binary can serve several
-        // compilers, telling them apart by the name it was invoked under.
-        let shim_name = format!(
-            "{}-{}",
-            language.shim_stem(),
-            variable.to_ascii_lowercase().replace(['.', '/'], "_")
-        );
-        let shim = shims_dir.join(shim_file_name(&shim_name));
-        link_path_shim(executable, &shim)?;
-        wrapped.push(TargetedCompiler {
-            variable,
-            shim_name,
-            shim,
-            real,
-        });
-    }
-    wrapped.sort_by(|left, right| left.variable.cmp(&right.variable));
-    Ok(wrapped)
-}
-
-/// Resolve a compiler a build named, rejecting anything that is not one path.
-fn resolve_named_compiler(value: &str, executable: &Path, shims: &Path) -> Option<PathBuf> {
-    let value = value.trim();
-    if value.is_empty() || value.split_whitespace().count() != 1 {
-        return None;
-    }
-    let candidate = Path::new(value);
-    let resolved = if candidate.is_absolute() {
-        candidate.is_file().then(|| candidate.to_path_buf())
-    } else {
-        resolve_on_path_excluding(value, executable, shims)
-    }?;
-    // A build already pointed at a shim -- this session's, or one an outer
-    // session left in the environment -- would otherwise be wrapped a second
-    // time, and the inner shim would exec itself. Comparing the directory as
-    // well as the binary catches the link that does not compare equal.
-    let inside_shims = resolved
-        .parent()
-        .is_some_and(|parent| canonical(parent) == canonical(shims));
-    (!inside_shims && !is_same_binary(&resolved, Some(executable))).then_some(resolved)
-}
-
-fn resolve_on_path(name: &str) -> Option<PathBuf> {
-    let path = std::env::var_os("PATH")?;
-    std::env::split_paths(&path)
-        .map(|directory| directory.join(name))
-        .find(|candidate| candidate.is_file())
-}
-
-/// A directory of compiler-named shims for builds that find their compiler on
-/// `PATH`, and the compilers those shims stand in for.
-pub struct PathShims {
-    pub directory: PathBuf,
-    pub compilers: BTreeMap<String, PathBuf>,
-}
-
-/// Install a shim under every plain compiler name that resolves on `PATH`.
-///
-/// Resolution happens here rather than in the shim so the whole build agrees
-/// on one compiler per name, and skips the running binary so a shim directory
-/// already on `PATH` cannot become its own compiler. A machine with none of
-/// the names simply gets no shim directory.
-///
-/// `directory` outlives the session on purpose. A configure step records the
-/// compiler it found by absolute path -- CMake writes it into `CMakeCache.txt`,
-/// autoconf into the generated makefiles -- and a session-local directory would
-/// leave that build permanently naming a path that no longer exists. A stable
-/// one keeps the recorded path resolvable, and keeps a later `cmake --build`
-/// running through the cache rather than around it. Nothing is added to any
-/// `PATH` but the one handed to a single `mbx exec` command.
-pub fn install_path_shims(directory: &Path) -> Result<Option<PathShims>> {
-    if !cfg!(unix) {
-        return Ok(None);
-    }
-    let executable = std::env::current_exe().wrap_err("failed to locate the running mbx binary")?;
-    std::fs::create_dir_all(directory)?;
-    let mut compilers = BTreeMap::new();
-    for (name, _) in PATH_SHIM_NAMES {
-        let destination = directory.join(name);
-        let Some(real) = resolve_on_path_excluding(name, &executable, directory) else {
-            continue;
-        };
-        // Belt and braces for the recursion the exclusion above prevents: a
-        // shim that stood in for itself would exec itself forever, so leave
-        // the name uncached rather than install that.
-        if canonical(&real) == canonical(&destination) {
-            debug!("{name} resolved to its own shim; leaving it uncached");
-            continue;
-        }
-        link_path_shim(&executable, &destination)?;
-        compilers.insert((*name).to_string(), real);
-    }
-    if compilers.is_empty() {
-        debug!("no C or C++ compilers were found on PATH; nothing to shim");
-        return Ok(None);
-    }
-    Ok(Some(PathShims {
-        directory: directory.to_path_buf(),
-        compilers,
-    }))
-}
-
-/// Point one compiler name at this binary, replacing a stale link in place.
-///
-/// A symlink rather than a hard link so an upgraded mbx is picked up without
-/// reinstalling, and because on macOS a hard link taken while the binary is
-/// replaced can be killed at exec.
-///
-/// The replacement goes through a temporary name and a rename because this
-/// directory is shared: another build may be executing the very name being
-/// replaced, and `rename` leaves it resolvable at every instant where removing
-/// and recreating would not.
-#[cfg(unix)]
-fn link_path_shim(executable: &Path, destination: &Path) -> Result<()> {
-    // Absolutized for the same reason [`symlink_shim`] does it: a symlink is
-    // resolved from the directory holding it, which here is the cache's shim
-    // directory and never the caller's. `current_exe` has been absolute on
-    // every platform mbx runs on, but its contract does not promise one, and a
-    // relative target would point inside the shim directory itself.
-    let target = std::path::absolute(executable)?;
-    if std::fs::read_link(destination).is_ok_and(|existing| existing == target) {
-        return Ok(());
-    }
-    let staging = destination.with_file_name(format!(
-        ".{}.{}.{}",
-        destination
-            .file_name()
-            .unwrap_or_default()
-            .to_string_lossy(),
-        std::process::id(),
-        SHIM_STAGING_NONCE.fetch_add(1, Ordering::Relaxed)
-    ));
-    let _ = std::fs::remove_file(&staging);
-    std::os::unix::fs::symlink(&target, &staging)?;
-    std::fs::rename(&staging, destination)
-        .wrap_err_with(|| format!("failed to install the shim {}", destination.display()))?;
-    Ok(())
-}
-
-#[cfg(not(unix))]
-fn link_path_shim(_executable: &Path, _destination: &Path) -> Result<()> {
-    eyre::bail!("PATH shims are not supported on this platform")
-}
-
-/// Find the real compiler `name` refers to, never a shim.
-///
-/// The shim directory is skipped by location, not by identity. Identity alone
-/// is not enough: a nested `mbx exec` sees that directory first on `PATH`, and
-/// a shim there may point at a *different* mbx binary -- an upgrade, or a build
-/// from another checkout -- which no inode comparison against the running one
-/// can recognize. Choosing it would pin a shim as the compiler and then relink
-/// it to the running binary, leaving it its own compiler and recursing until
-/// the process table fills. Nothing mbx puts in that directory is ever a real
-/// compiler, so the directory itself is the honest thing to exclude.
-///
-/// The identity check stays for the rest of `PATH`, where a copy of mbx under a
-/// compiler's name can still turn up outside any directory mbx owns.
-fn resolve_on_path_excluding(name: &str, this_binary: &Path, shims: &Path) -> Option<PathBuf> {
-    resolve_in_path(&std::env::var_os("PATH")?, name, this_binary, shims)
-}
-
-/// The search itself, over a `PATH` the caller supplies.
-///
-/// Separated from the environment so a test can hand it one: `PATH` is process
-/// global, and cargo runs these tests on a thread pool, so setting it here
-/// would race every other test that reads one -- `which` in the linker tests
-/// among them.
-fn resolve_in_path(path: &OsStr, name: &str, this_binary: &Path, shims: &Path) -> Option<PathBuf> {
-    let shims = canonical(shims);
-    std::env::split_paths(path)
-        .filter(|directory| canonical(directory) != shims)
-        .map(|directory| directory.join(name))
-        .find(|candidate| candidate.is_file() && !is_same_binary(candidate, Some(this_binary)))
-}
-
-fn canonical(path: &Path) -> PathBuf {
-    std::fs::canonicalize(path).unwrap_or_else(|_| path.to_path_buf())
-}
-
-fn install_session_shim(session_dir: &Path) -> Result<PathBuf> {
-    let executable = std::env::current_exe().wrap_err("failed to locate the running mbx binary")?;
-    install_shim(&executable, session_dir, ShimLink::Tracking)
-}
-
-/// How an installed shim refers to the mbx binary behind it.
-#[derive(Clone, Copy, Debug, PartialEq, Eq)]
-pub enum ShimLink {
-    /// A symlink: the shim is whatever binary that path holds when it runs.
-    Tracking,
-    /// A hard link, or a copy where the filesystem cannot link: the bytes that
-    /// were there on the day the shim was installed.
-    Pinned,
-}
-
-/// File name a shim with this stem is installed under.
-pub fn shim_file_name(stem: &str) -> String {
-    if cfg!(windows) {
-        format!("{stem}.exe")
-    } else {
-        stem.into()
-    }
-}
-
-/// Install a rustc shim into `directory` as a link to `executable`.
-///
-/// The shim must be the same binary as the agent -- the handshake requires an
-/// exact version match -- so it is a link to the running binary rather than an
-/// independently built copy.
-///
-/// Which kind of link matters more than it looks. On macOS, exec of a path that
-/// `link(2)` created moments ago is intermittently killed outright: SIGKILL, no
-/// output, no crash report, nothing in the system log. Measured on macOS 26.6
-/// (Apple silicon) at about one exec in eight hundred under heavy parallel
-/// load, against a binary nothing was writing, and far more readily for a large
-/// image -- the 50 MB debug build of mbx died where a 500 KB one never did,
-/// which is the shape a race in per-page signature validation would have. The
-/// window closes about half a second after the link appears, and the same path
-/// then runs fine forever. Exec of the original path never fails, and neither
-/// does exec of a symlink to it -- the kernel resolves that to a file whose
-/// signature it validated long ago. Reading the new link through first,
-/// fsyncing it and its directory, and taking the first exec ourselves to spend
-/// the race were all tried, and none of them helped.
-///
-/// So [`ShimLink::Tracking`] is for the session shim, which cargo execs a few
-/// milliseconds after it is installed to ask `rustc -vV` what compiler it has.
-/// [`ShimLink::Pinned`] is for the plain-cargo wrapper `mbx setup` installs,
-/// where nothing execs the shim for as long as it takes to type another command
-/// and a hard link keeps working even if the binary it was made from is deleted.
-///
-/// The one thing tracking gives up: replace the mbx binary underneath a running
-/// build and the session shim follows it, so cargo either execs nothing (the
-/// path is gone, and the build stops with a plain error) or execs a version the
-/// agent will not shake hands with, which bypasses the cache for the rest of the
-/// build. Both are loud and self-inflicted, unlike the kill they replace.
-pub fn install_shim(executable: &Path, directory: &Path, link: ShimLink) -> Result<PathBuf> {
-    install_shim_named(executable, directory, RUSTC_SHIM_STEM, link)
-}
-
-/// Install a shim for one compiler into `directory` as a link to `executable`.
-///
-/// Every shim is the same binary under a different name; the name is what tells
-/// it which compiler it stands in for. Everything [`install_shim`] documents
-/// about which kind of link is used applies here too.
-pub fn install_shim_named(
-    executable: &Path,
-    directory: &Path,
-    stem: &str,
-    link: ShimLink,
-) -> Result<PathBuf> {
-    let shim = directory.join(shim_file_name(stem));
-    let _ = std::fs::remove_file(&shim);
-    if link == ShimLink::Tracking && symlink_shim(executable, &shim) {
-        return Ok(shim);
-    }
-    if let Err(link_error) = std::fs::hard_link(executable, &shim) {
-        std::fs::copy(executable, &shim).wrap_err_with(|| {
-            format!("failed to install the {stem} shim by hard link ({link_error}) or copy")
-        })?;
-    }
-    #[cfg(unix)]
-    {
-        use std::os::unix::fs::PermissionsExt as _;
-
-        // Some filesystems do not retain executable permissions when the
-        // cross-device fallback copies the running binary. Cargo must be able
-        // to invoke the installed wrapper directly.
-        let mut permissions = std::fs::metadata(&shim)?.permissions();
-        permissions.set_mode(permissions.mode() | 0o100);
-        std::fs::set_permissions(&shim, permissions)?;
-    }
-    Ok(shim)
-}
-
-/// Point `shim` at `executable` by symlink, reporting whether that worked.
-///
-/// A failure is not an error: the hard link the caller falls back to is what
-/// every shim was before, so the only thing lost is the race described on
-/// [`install_shim`].
-///
-/// The target is absolutized first, because the two kinds of link do not read a
-/// relative one the same way: `link(2)` and `copy` resolve it from the caller's
-/// working directory, while a symlink resolves it from the shim's own directory
-/// -- a temporary one that shares nothing with the caller's. Resolving it here
-/// gives the argument one meaning. Absolutizing rather than declining, because a
-/// relative target must still get a symlink: `current_exe()` has been absolute
-/// on every platform mbx runs on, but nothing in its contract promises that, and
-/// a shim that quietly stopped tracking would put the kill back.
-///
-/// A target that is not there gets no symlink at all. A hard link and a copy
-/// both refuse one, and a symlink would instead name it and leave cargo to
-/// discover the dangling wrapper mid-build.
-#[cfg(unix)]
-fn symlink_shim(executable: &Path, shim: &Path) -> bool {
-    let Ok(target) = std::path::absolute(executable) else {
-        return false;
-    };
-    target.exists() && std::os::unix::fs::symlink(&target, shim).is_ok()
-}
-
-/// Windows has no shim symlinks: creating one needs a privilege ordinary
-/// accounts lack, and the code-signature race they exist to avoid is a macOS
-/// kernel behaviour with no Windows counterpart.
-#[cfg(windows)]
-fn symlink_shim(_executable: &Path, _shim: &Path) -> bool {
-    false
-}
-
-#[cfg(unix)]
-async fn spawn_server(
-    session_dir: &Path,
-    agent: CacheAgent,
-    mut shutdown: oneshot::Receiver<()>,
-) -> Result<(String, JoinHandle<Result<()>>)> {
-    use std::os::unix::fs::PermissionsExt as _;
-
-    let socket = session_dir.join("cache-agent.sock");
-    let listener = tokio::net::UnixListener::bind(&socket)?;
-    std::fs::set_permissions(&socket, std::fs::Permissions::from_mode(0o600))?;
-    let endpoint = socket.to_string_lossy().into_owned();
-    let server = tokio::spawn(async move {
-        let _cleanup = SocketCleanup(socket);
-        let mut connections = tokio::task::JoinSet::new();
-        let outcome = loop {
-            tokio::select! {
-                accepted = listener.accept() => {
-                    let (stream, _) = match accepted {
-                        Ok(accepted) => accepted,
-                        Err(error) => break Err(eyre::Report::from(error)),
-                    };
-                    let agent = agent.clone();
-                    connections.spawn(async move {
-                        if let Err(error) = agent.handle_connection(stream).await {
-                            debug!("cache agent connection failed: {error}");
-                        }
-                    });
-                    // A build makes one connection per compilation, so finished
-                    // ones are reaped as they go rather than held until shutdown.
-                    while connections.try_join_next().is_some() {}
-                }
-                _ = &mut shutdown => break Ok(()),
-            }
-        };
-        // A connection answers requests, and a store request queues an upload,
-        // so the session has not stopped accepting work until these are done.
-        while connections.join_next().await.is_some() {}
-        outcome
-    });
-    Ok((endpoint, server))
-}
-
-#[cfg(unix)]
-struct SocketCleanup(PathBuf);
-
-#[cfg(unix)]
-impl Drop for SocketCleanup {
-    fn drop(&mut self) {
-        let _ = std::fs::remove_file(&self.0);
-    }
-}
-
-#[cfg(windows)]
-async fn spawn_server(
-    _session_dir: &Path,
-    agent: CacheAgent,
-    mut shutdown: oneshot::Receiver<()>,
-) -> Result<(String, JoinHandle<Result<()>>)> {
-    let endpoint = format!(
-        r"\\.\pipe\mbx-cache-{}-{}",
-        std::process::id(),
-        crate::util::random_string(12)
-    );
-    let first_server = create_named_pipe(&endpoint, true)?;
-    let server_endpoint = endpoint.clone();
-    let server = tokio::spawn(async move {
-        let mut next_server = Some(first_server);
-        let mut connections = tokio::task::JoinSet::new();
-        let outcome = loop {
-            let pipe = next_server
-                .take()
-                .expect("the next named-pipe server is always prepared");
-            tokio::select! {
-                connected = pipe.connect() => {
-                    if let Err(error) = connected {
-                        break Err(eyre::Report::from(error));
-                    }
-                    match create_named_pipe(&server_endpoint, false) {
-                        Ok(prepared) => next_server = Some(prepared),
-                        Err(error) => break Err(error),
-                    }
-                    let agent = agent.clone();
-                    connections.spawn(async move {
-                        if let Err(error) = agent.handle_connection(pipe).await {
-                            debug!("cache agent connection failed: {error}");
-                        }
-                    });
-                    // A build makes one connection per compilation, so finished
-                    // ones are reaped as they go rather than held until shutdown.
-                    while connections.try_join_next().is_some() {}
-                }
-                _ = &mut shutdown => break Ok(()),
-            }
-        };
-        // A connection answers requests, and a store request queues an upload,
-        // so the session has not stopped accepting work until these are done.
-        while connections.join_next().await.is_some() {}
-        outcome
-    });
-    Ok((endpoint, server))
-}
-
-#[cfg(windows)]
-fn create_named_pipe(
-    endpoint: &str,
-    first_pipe_instance: bool,
-) -> Result<tokio::net::windows::named_pipe::NamedPipeServer> {
-    use std::mem::size_of;
-    use windows_sys::Win32::Security::SECURITY_ATTRIBUTES;
-
-    let security = CurrentUserSecurityDescriptor::new()?;
-    let mut attributes = SECURITY_ATTRIBUTES {
-        nLength: size_of::<SECURITY_ATTRIBUTES>() as u32,
-        lpSecurityDescriptor: security.0,
-        bInheritHandle: 0,
-    };
-    let mut options = tokio::net::windows::named_pipe::ServerOptions::new();
-    options.first_pipe_instance(first_pipe_instance);
-    // SAFETY: `attributes` and its owned security descriptor remain valid for
-    // the duration of CreateNamedPipeW, and the handle is not inheritable.
-    unsafe {
-        options
-            .create_with_security_attributes_raw(endpoint, (&raw mut attributes).cast())
-            .wrap_err("failed to create the current-user-only cache pipe")
-    }
-}
-
-#[cfg(windows)]
-struct CurrentUserSecurityDescriptor(windows_sys::Win32::Security::PSECURITY_DESCRIPTOR);
-
-#[cfg(windows)]
-impl CurrentUserSecurityDescriptor {
-    fn new() -> Result<Self> {
-        use std::mem::size_of;
-        use std::ptr::null_mut;
-        use windows_sys::Win32::Foundation::HANDLE;
-        use windows_sys::Win32::Security::Authorization::{
-            ConvertSidToStringSidW, ConvertStringSecurityDescriptorToSecurityDescriptorW,
-            SDDL_REVISION_1,
-        };
-        use windows_sys::Win32::Security::{
-            GetTokenInformation, TOKEN_QUERY, TOKEN_USER, TokenUser,
-        };
-        use windows_sys::Win32::System::Threading::{GetCurrentProcess, OpenProcessToken};
-
-        let mut token: HANDLE = null_mut();
-        // SAFETY: `token` is a valid out pointer and the process pseudo-handle
-        // is always valid in the calling process.
-        if unsafe { OpenProcessToken(GetCurrentProcess(), TOKEN_QUERY, &raw mut token) } == 0 {
-            return Err(std::io::Error::last_os_error())
-                .wrap_err("failed to open the current process token");
-        }
-        let token = OwnedWindowsHandle(token);
-
-        let mut required = 0;
-        // The first call intentionally obtains the required buffer size.
-        // SAFETY: a null information pointer is required for this size query.
-        unsafe {
-            GetTokenInformation(token.0, TokenUser, null_mut(), 0, &raw mut required);
-        }
-        if required == 0 {
-            return Err(std::io::Error::last_os_error())
-                .wrap_err("failed to size the current process user token");
-        }
-        let word_count = (required as usize).div_ceil(size_of::<usize>());
-        let mut token_information = vec![0usize; word_count];
-        // SAFETY: the aligned buffer is at least `required` bytes long and the
-        // API initializes it as TOKEN_USER on success.
-        if unsafe {
-            GetTokenInformation(
-                token.0,
-                TokenUser,
-                token_information.as_mut_ptr().cast(),
-                required,
-                &raw mut required,
-            )
-        } == 0
-        {
-            return Err(std::io::Error::last_os_error())
-                .wrap_err("failed to read the current process user token");
-        }
-        // SAFETY: GetTokenInformation successfully initialized the aligned
-        // buffer as TOKEN_USER and its SID remains owned by that buffer.
-        let user_sid = unsafe {
-            (*(token_information.as_ptr().cast::<TOKEN_USER>()))
-                .User
-                .Sid
-        };
-        let mut sid_string = null_mut();
-        // SAFETY: `user_sid` points into the live token-information buffer and
-        // `sid_string` is a valid out pointer.
-        if unsafe { ConvertSidToStringSidW(user_sid, &raw mut sid_string) } == 0 {
-            return Err(std::io::Error::last_os_error())
-                .wrap_err("failed to format the current process user SID");
-        }
-        let sid_string = LocalWindowsAllocation(sid_string.cast());
-        // SAFETY: ConvertSidToStringSidW returned a valid NUL-terminated string
-        // whose allocation remains live through `sid_string`.
-        let sid = unsafe { nul_terminated_wide(sid_string.0.cast()) };
-
-        let mut sddl: Vec<u16> = "D:P(A;;GA;;;".encode_utf16().collect();
-        sddl.extend_from_slice(sid);
-        sddl.extend([')' as u16, 0]);
-        let mut descriptor = null_mut();
-        // SAFETY: the SDDL is NUL-terminated, `descriptor` is a valid out
-        // pointer, and the returned allocation is owned by LocalFree.
-        if unsafe {
-            ConvertStringSecurityDescriptorToSecurityDescriptorW(
-                sddl.as_ptr(),
-                SDDL_REVISION_1,
-                &raw mut descriptor,
-                null_mut(),
-            )
-        } == 0
-        {
-            return Err(std::io::Error::last_os_error())
-                .wrap_err("failed to restrict the cache pipe to the current user");
-        }
-        drop(token);
-        drop(sid_string);
-        debug_assert!(!descriptor.is_null());
-        Ok(Self(descriptor))
-    }
-}
-
-#[cfg(windows)]
-impl Drop for CurrentUserSecurityDescriptor {
-    fn drop(&mut self) {
-        // SAFETY: this allocation came from
-        // ConvertStringSecurityDescriptorToSecurityDescriptorW.
-        unsafe {
-            windows_sys::Win32::Foundation::LocalFree(self.0.cast());
-        }
-    }
-}
-
-#[cfg(windows)]
-struct OwnedWindowsHandle(windows_sys::Win32::Foundation::HANDLE);
-
-#[cfg(windows)]
-impl Drop for OwnedWindowsHandle {
-    fn drop(&mut self) {
-        // SAFETY: this handle came from OpenProcessToken and is owned here.
-        unsafe {
-            windows_sys::Win32::Foundation::CloseHandle(self.0);
-        }
-    }
-}
-
-#[cfg(windows)]
-struct LocalWindowsAllocation(windows_sys::Win32::Foundation::HLOCAL);
-
-#[cfg(windows)]
-impl Drop for LocalWindowsAllocation {
-    fn drop(&mut self) {
-        // SAFETY: this allocation came from a Win32 API documented to use
-        // LocalAlloc and has not previously been freed.
-        unsafe {
-            windows_sys::Win32::Foundation::LocalFree(self.0);
-        }
-    }
-}
-
-#[cfg(windows)]
-unsafe fn nul_terminated_wide<'a>(value: *const u16) -> &'a [u16] {
-    let mut length = 0;
-    // SAFETY: the caller guarantees that `value` points to a valid
-    // NUL-terminated UTF-16 string.
-    while unsafe { *value.add(length) } != 0 {
-        length += 1;
-    }
-    // SAFETY: the scan above established the initialized string length.
-    unsafe { std::slice::from_raw_parts(value, length) }
 }
 
 /// Whether this process was invoked as the rustc shim.
@@ -2349,242 +1228,6 @@ pub(crate) fn request_agent(requests: &[AgentRequest]) -> Result<Vec<AgentRespon
 /// compiler owns, which is what [`report_shim_warning`] exists to avoid.
 fn session_socket() -> Option<OsString> {
     std::env::var_os(SOCKET_ENV).filter(|socket| !socket.is_empty())
-}
-
-/// Serve a persistent Cargo wrapper directly from the local store.
-///
-/// There is deliberately no remote access here: the short-lived process has
-/// no trusted session policy or opportunity to batch transfers. Prediction
-/// manifests are still persisted after every successful record so the many
-/// wrapper processes in one Cargo build, and later builds, share what they
-/// learn without a daemon.
-fn request_standalone_agent(requests: &[AgentRequest]) -> Result<Vec<AgentResponse>> {
-    thread_local! {
-        static STANDALONE_AGENT: RefCell<Option<StandaloneAgent>> = const { RefCell::new(None) };
-    }
-
-    STANDALONE_AGENT.with(|slot| {
-        let mut slot = slot.borrow_mut();
-        if slot.is_none() {
-            let config = Config::load()?;
-            let runtime = tokio::runtime::Builder::new_current_thread()
-                .enable_all()
-                .build()?;
-            *slot = Some(StandaloneAgent {
-                agent: CacheAgent::new(config.store_dir(), VERSION),
-                runtime,
-                runs: BTreeMap::new(),
-            });
-        }
-        let standalone = slot.as_mut().expect("standalone agent was initialized");
-        let StandaloneAgent {
-            agent,
-            runtime,
-            runs,
-        } = standalone;
-        runtime.block_on(async {
-            let mut responses = Vec::with_capacity(requests.len());
-            for request in requests.iter().cloned() {
-                let (request, run_to_commit) = match request {
-                    AgentRequest::FindActionPrediction { task, invocation } => {
-                        let run = match runs.get(&task) {
-                            Some(run) => run.clone(),
-                            None => {
-                                let run = agent.begin_task(&task).await?;
-                                runs.insert(task.clone(), run.clone());
-                                run
-                            }
-                        };
-                        (
-                            AgentRequest::FindActionPrediction {
-                                task: run,
-                                invocation,
-                            },
-                            None,
-                        )
-                    }
-                    AgentRequest::RecordActionPrediction { task, prediction } => {
-                        let run = match runs.remove(&task) {
-                            Some(run) => run,
-                            None => agent.begin_task(&task).await?,
-                        };
-                        (
-                            AgentRequest::RecordActionPrediction {
-                                task: run.clone(),
-                                prediction,
-                            },
-                            Some(run),
-                        )
-                    }
-                    request => (request, None),
-                };
-                let response = agent
-                    .handle_requests(std::iter::once(request))
-                    .await
-                    .into_iter()
-                    .next()
-                    .expect("one request returns one response");
-                let succeeded = !matches!(response, AgentResponse::Error { .. });
-                responses.push(response);
-                if succeeded && let Some(run) = run_to_commit {
-                    agent.commit_task(&run).await?;
-                }
-            }
-            Ok(responses)
-        })
-    })
-}
-
-struct StandaloneAgent {
-    agent: CacheAgent,
-    runtime: tokio::runtime::Runtime,
-    runs: BTreeMap<String, String>,
-}
-
-#[cfg(unix)]
-fn request_agent_at(socket: &OsString, requests: &[AgentRequest]) -> Result<Vec<AgentResponse>> {
-    // One compilation asks the agent several separate questions, and paying a
-    // connect plus a handshake round trip for each adds up across a warm
-    // build. The protocol is strict request-response, so a connection held for
-    // the life of this shim process serves them all.
-    thread_local! {
-        static CONNECTION: RefCell<Option<(OsString, std::os::unix::net::UnixStream)>> =
-            const { RefCell::new(None) };
-    }
-    CONNECTION.with(|slot| {
-        let mut slot = slot.borrow_mut();
-        if slot.as_ref().is_none_or(|(known, _)| known != socket) {
-            let mut stream = std::os::unix::net::UnixStream::connect(Path::new(socket))
-                .wrap_err("failed to connect to the cache session")?;
-            sync_handshake(&mut stream)?;
-            *slot = Some((socket.clone(), stream));
-        }
-        let (_, stream) = slot.as_mut().expect("the connection was just established");
-        let responses = requests
-            .iter()
-            .map(|request| sync_request(stream, request))
-            .collect::<Result<Vec<_>>>();
-        if responses.is_err() {
-            // A failed exchange leaves the stream in an unknown protocol
-            // state, so the next request starts over on a fresh connection.
-            *slot = None;
-        }
-        responses
-    })
-}
-
-#[cfg(windows)]
-fn request_agent_at(socket: &OsString, requests: &[AgentRequest]) -> Result<Vec<AgentResponse>> {
-    let endpoint = socket.to_string_lossy().into_owned();
-    tokio::runtime::Builder::new_current_thread()
-        .enable_all()
-        .build()?
-        .block_on(async move {
-            use tokio::io::{AsyncBufReadExt, AsyncWriteExt};
-            let mut stream = connect_pipe(&endpoint).await?;
-            let request = AgentRequest::Hello {
-                protocol: AGENT_PROTOCOL_VERSION,
-                client_version: VERSION.into(),
-            };
-            let mut encoded = serde_json::to_vec(&request)?;
-            encoded.push(b'\n');
-            stream.write_all(&encoded).await?;
-            stream.flush().await?;
-            let mut response = String::new();
-            tokio::io::BufReader::new(&mut stream)
-                .read_line(&mut response)
-                .await?;
-            validate_handshake_response(&response)?;
-            let mut responses = Vec::with_capacity(requests.len());
-            for request in requests {
-                let mut encoded = serde_json::to_vec(request)?;
-                encoded.push(b'\n');
-                stream.write_all(&encoded).await?;
-                stream.flush().await?;
-                let mut response = String::new();
-                tokio::io::BufReader::new(&mut stream)
-                    .read_line(&mut response)
-                    .await?;
-                responses.push(serde_json::from_str(&response)?);
-            }
-            Ok(responses)
-        })
-}
-
-/// Open the agent's pipe, waiting out the instances that are already busy.
-///
-/// The agent keeps one spare instance and creates the next only after accepting,
-/// so the rustc processes cargo runs in parallel routinely arrive to a busy
-/// pipe. Without this they would fall back to uncached compiles.
-#[cfg(windows)]
-async fn connect_pipe(endpoint: &str) -> Result<tokio::net::windows::named_pipe::NamedPipeClient> {
-    use std::time::Instant;
-    use windows_sys::Win32::Foundation::ERROR_PIPE_BUSY;
-
-    const RETRY_DEADLINE: std::time::Duration = std::time::Duration::from_secs(10);
-    const RETRY_DELAY: std::time::Duration = std::time::Duration::from_millis(5);
-
-    let deadline = Instant::now() + RETRY_DEADLINE;
-    loop {
-        match tokio::net::windows::named_pipe::ClientOptions::new().open(endpoint) {
-            Ok(client) => return Ok(client),
-            Err(error)
-                if error.raw_os_error() == Some(ERROR_PIPE_BUSY as i32)
-                    && Instant::now() < deadline => {}
-            Err(error) => {
-                return Err(error).wrap_err("failed to connect to the cache session");
-            }
-        }
-        tokio::time::sleep(RETRY_DELAY).await;
-    }
-}
-
-#[cfg(unix)]
-fn sync_handshake(stream: &mut (impl std::io::Read + Write)) -> Result<()> {
-    let request = AgentRequest::Hello {
-        protocol: AGENT_PROTOCOL_VERSION,
-        client_version: VERSION.into(),
-    };
-    write_request_line(stream, &request)?;
-    let mut response = String::new();
-    BufReader::new(&mut *stream).read_line(&mut response)?;
-    validate_handshake_response(&response)
-}
-
-#[cfg(unix)]
-fn sync_request(
-    stream: &mut (impl std::io::Read + Write),
-    request: &AgentRequest,
-) -> Result<AgentResponse> {
-    write_request_line(stream, request)?;
-    let mut response = String::new();
-    BufReader::new(&mut *stream).read_line(&mut response)?;
-    Ok(serde_json::from_str(&response)?)
-}
-
-/// Send one request as a single write.
-///
-/// Serializing straight onto the socket emits a syscall per JSON fragment,
-/// and a prediction payload has thousands of them, each waking the agent's
-/// reader. The encoded line is assembled first so the kernel sees it once.
-#[cfg(unix)]
-fn write_request_line(stream: &mut impl Write, request: &AgentRequest) -> Result<()> {
-    let mut encoded = serde_json::to_vec(request)?;
-    encoded.push(b'\n');
-    stream.write_all(&encoded)?;
-    stream.flush()?;
-    Ok(())
-}
-
-fn validate_handshake_response(response: &str) -> Result<()> {
-    match serde_json::from_str(response)? {
-        AgentResponse::Hello {
-            protocol,
-            agent_version,
-        } if protocol == AGENT_PROTOCOL_VERSION && agent_version == VERSION => Ok(()),
-        AgentResponse::Error { message } => bail!(message),
-        _ => bail!("the cache agent returned an incompatible handshake"),
-    }
 }
 
 #[cfg(test)]
